pyproject.toml
[tool.poetry]
name = "paas-devops"
version = "0.1.0"
description = "Generic APIS project for auto devops"
authors = ["Matthias Schlögl <m.schloegl@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "~3.8"
whitenoise = "^5.2.0"
sentry-sdk = "^0.19.5"
dj-database-url = "^0.5.0"
gunicorn = "^20.0.4"
acdh-django-charts = "^0.5.3"
django-allow-cidr = "^0.3.1"
apis-core = "~0.16.33"
mysqlclient = "^2.0.3"
django-extensions = "^3.1.3"
django-csp = "^3.7"
apis-highlighter = "^0.8.18"
apis-bibsonomy = "^0.2.3"
viecpro-vis = "^0.1.9"
django-haystack = "^3.1.1"
pysolr = "^3.9.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.0"
pylint = "^2.6.0"
dephell = {extras = ["full"], version = "^0.8.3"}
mistune = "<2.0.0"
jupyter = "^1.0.0"
jupyter-tabnine = "^1.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
